Feat/add order tests

+++ src/useCases/order.ts
@@ -1,32 +1,57 @@
-import { Order } from '../domain/entities/order';
-import { OrderRepository } from '../domain/interface/orderRepository';
+import { Order } from '../domain/entities/order'
+import { OrderRepository } from '../domain/interface/orderRepository'
+import { ClientRepository } from '../domain/interface/clientRepository'
 
 export class OrderUseCase {
-  constructor(private orderRepository: OrderRepository) { }
+    constructor(
+        private orderRepository: OrderRepository,
+        private clientRepository: ClientRepository
+    ) {}
 
-  async getOrder(id: string): Promise<Order | null> {
-    return await this.orderRepository.getOrder(id);
-  }
+    async getOrder(id: string): Promise<Order | null> {
+        return await this.orderRepository.getOrder(id)
+    }
 
-  async createOrder(order: Order): Promise<void> {
-    // validar se o produto existe.
-    return await this.orderRepository.createOrder(order);
-  } 
+    async createOrder(order: Order): Promise<void> {
+        if (order.idOrder) {
+            const existingOrder = await this.getOrder(order.idOrder)
+            if (existingOrder) {
+                throw new Error('Order already exist')
+            }
+        }
 
-  async updateOrder(id: string, order: Order) {
-    return await this.orderRepository.updateOrder(id, order);
-  }
-  
-  async deleteOrder(id: string) {
-    return await this.orderRepository.deleteOrder(id);
-  }
+        if (order.idClient) {
+            const existingClient = await this.clientRepository.findById(
+                order.idClient
+            )
+            if (!existingClient) {
+                throw new Error('Client does not exist')
+            }
+        }
+        return await this.orderRepository.createOrder(order)
+    }
 
-  async updateOrderStatus(id: string, status: string): Promise<Order | null> {
-    return await this.orderRepository.updateOrderStatus(id, status)
-  }
+    async updateOrder(id: string, order: Order) {
+        const existingOrder = await this.getOrder(id)
+        if (!existingOrder) {
+            throw new Error('Order does not exist')
+        }
+        return await this.orderRepository.updateOrder(id, order)
+    }
 
-  async listOrders( ): Promise<Order[]> {
-    return await this.orderRepository.listOrders();
-  }
+    async deleteOrder(id: string) {
+        const existingOrder = await this.getOrder(id)
+        if (!existingOrder) {
+            throw new Error('Order does not exist')
+        }
+        return await this.orderRepository.deleteOrder(id)
+    }
 
+    async updateOrderStatus(id: string, status: string): Promise<Order | null> {
+        return await this.orderRepository.updateOrderStatus(id, status)
+    }
+
+    async listOrders(): Promise<Order[]> {
+        return await this.orderRepository.listOrders()
+    }
 }
